requirements.txt
flask
flask_socketio
Pillow